+++ qadence/backends/utils.py
@@ -231,52 +231,3 @@ def dydxx(
         ),
         values[op.param_name],
     )
-
-
-def _torch_derivative(
-    ufa: Callable, x: torch.Tensor, derivative_indices: tuple[int, ...]
-) -> torch.Tensor:
-    y = ufa(x)
-    for idx in derivative_indices:
-        out = torch.autograd.grad(y, x, torch.ones_like(y), create_graph=True)[0]
-        y = out[:, idx]
-    return y.reshape(-1, 1)
-
-
-def derivative(ufa: torch.nn.Module, x: Tensor, derivative_indices: tuple[int, ...]) -> Tensor:
-    """Compute derivatives w.r.t.
-
-    inputs of a UFA with a single output. The
-    `derivative_indices` specify which derivative(s) are computed.  E.g.
-    `derivative_indices=(1,2)` would compute the a second order derivative w.r.t
-    to the indices `1` and `2` of the input tensor.
-
-    Arguments:
-        ufa: The model for which we want to compute the derivative.
-        x (Tensor): (batch_size, input_size) input tensor.
-        derivative_indices (tuple): Define which derivatives to compute.
-
-    Examples:
-    If we create a UFA with three inputs and denote the first, second, and third
-    input with `x`, `y`, and `z` we can compute the following derivatives w.r.t
-    to those inputs:
-    ```py exec="on" source="material-block"
-    import torch
-    from qadence import QNN
-    from qadence.backends.utils import derivative
-
-    f = MLP([3,3,1])
-    inputs = torch.rand(5,3,requires_grad=True)
-
-    # df_dx
-    derivative(f, inputs, (0,))
-
-    # d2f_dydz
-    derivative(f, inputs, (1,2))
-
-    # d3fdy2dx
-    derivative(f, inputs, (1,1,0))
-    ```
-    """
-    assert ufa.out_features == 1, "Can only call `derivative` on models with 1D output."
-    return ufa._derivative(x, derivative_indices)
